+++ logica-js-projeto_inicial/app.js
@@ -1,29 +1,23 @@
-// let titulo = document.querySelector("h1");
-// titulo.innerHTML = "Jogo do número secreto";
-
-// let paragrafo = document.querySelector("p");
-// paragrafo.innerHTML = "Escolha um número entre 1 e 10:";
-
 let listaDeNumerosSorteados = [];
-let numeroMaximo = 10;
+let numeroMaximo = 100;
 let numeroSecreto = gerarNumeroAleatorio();
 let tentativas = 1;
 
 function exibirTextoNaTela(tag, texto) {
     let campo = document.querySelector(tag);
     campo.innerHTML = texto;
+    responsiveVoice.speak(texto, "Brazilian Portuguese Female", {rate: 1.2});
 }
 
 function exibirMensagemInicial() {
     exibirTextoNaTela("h1", "Jogo do número secreto");
-    exibirTextoNaTela("p", "Escolha um número entre 1 e 10:");
+    exibirTextoNaTela("p", "Escolha um número entre 1 e 100:");
 }
 
 exibirMensagemInicial();
 
 function verificarChute() {
     let chute = document.querySelector("input").value;
-    // console.log(chute == numeroSecreto);
  
     if(chute == numeroSecreto) {
         exibirTextoNaTela("h1", "Acertou!");